requirements.txt
playwright
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
pytest

# Qase pytest integration
qase-pytest
